changing return type inside the program

+++ 0x1E-search_algorithms/0-linear.c
@@ -19,7 +19,7 @@ int linear_search(int *array, size_t size, int value)
 	{
 		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
-			return (i);
+			return ((size_t)i);
 		i++;
 	}
